apps/ai-workers/thumbnail-generator/requirements.txt
# Core dependencies
boto3==1.34.162
python-dotenv==1.0.1
pydantic==2.8.2
requests==2.32.3

# Image processing dependencies
pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==1.26.4

# AI/ML dependencies
torch==2.4.0
torchvision==0.19.0
ultralytics==8.0.196
diffusers==0.30.0

# AWS Lambda dependencies
aws-lambda-powertools==2.38.0

# Video processing
ffmpeg-python==0.2.0

# Logging and monitoring
structlog==24.4.0

# Testing
pytest==8.3.2
pytest-asyncio==0.23.8
moto==5.0.12

# Development
black==24.8.0
flake8==7.1.1
mypy==1.11.2
